+++ CountlyTests/CountlyTests-Bridging-Header.h
@@ -3,3 +3,6 @@
 #import "CountlyViewTrackingInternal.h"
 #import "CountlyHealthTracker.h"
 #import "CountlyContentBuilderInternal.h"
+#import "CountlyWebViewManager.h"
+#import "CountlyWebViewManager+Tests.h"
+#import "PassThroughBackgroundView.h"